fix minor changes and review comments
