+++ pkg/gofr/cmd/terminal/printers.go
@@ -4,14 +4,14 @@ import (
 	"fmt"
 )
 
-func (o *Output) Printf(format string, args ...interface{}) {
+func (o *output) Printf(format string, args ...interface{}) {
 	fmt.Fprintf(o.out, format, args...)
 }
 
-func (o *Output) Print(messages ...interface{}) {
+func (o *output) Print(messages ...interface{}) {
 	fmt.Fprint(o.out, messages...)
 }
 
-func (o *Output) Println(messages ...interface{}) {
+func (o *output) Println(messages ...interface{}) {
 	fmt.Fprintln(o.out, messages...)
 }
